pyproject.toml
[project]
authors = [
  {name = "Saito Tsutomu", email = "tsutomu.saito@moai-lab.jp"},
]
dependencies = [
  "nicegui-book-list>=0.1.1",
]
description = "niceguiのpytestのサンプル"
name = "nicegui-pytest"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.urls]
homepage = "https://github.com/SaitoTsutomu/nicegui-pytest"

[project.scripts]
book-list = "nicegui_pytest:main"

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pytest-selenium>=4.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
files = "src"
ignore_missing_imports = true

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
ignore = ["CPY", "D1", "D2", "D4", "DOC201", "E501", "PD901"]
preview = true
select = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]
